requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
gunicorn==20.1.0